@@ -1,12 +1,12 @@
 {
   "name": "mui-phone-number",
-  "version": "2.2.11",
+  "version": "3.0.0",
   "description": "A material-ui v5+ react component to format phone numbers. Based on react-phone-input-2",
   "main": "dist/index.js",
   "types": "index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack-dev-server --config webpack.config.dev.js --hot --progress --mode development",
+    "dev": "webpack serve --config webpack.config.dev.js --hot --progress --mode development",
     "build": "webpack --config webpack.config.build.js --progress --mode production"
   },
   "repository": {
@@ -26,50 +26,46 @@
   },
   "homepage": "https://github.com/phrensoua/mui-phone-number",
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.5.4",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@types/react": "^17.0.8",
-    "babel-eslint": "^10.0.3",
+    "@babel/core": "^7.15.8",
+    "@babel/plugin-proposal-class-properties": "^7.14.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.15.6",
+    "@babel/preset-env": "^7.15.8",
+    "@babel/preset-react": "^7.14.5",
+    "@types/react": "^17.0.28",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "babel-plugin-lodash": "^3.3.4",
-    "css-loader": "^3.6.0",
-    "eslint": "^6.2.2",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.3",
-    "file-loader": "^4.3.0",
-    "html-webpack-plugin": "^4.0.0",
-    "less": "^3.9.0",
-    "less-loader": "^5.0.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "css-loader": "^6.4.0",
+    "eslint": "^8.0.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.25.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.26.1",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.3.2",
+    "less": "^4.1.2",
+    "less-loader": "^10.1.0",
+    "optimize-css-assets-webpack-plugin": "^6.0.1",
     "path": "^0.12.7",
     "react-hot-loader": "^4.13.0",
-    "style-loader": "^0.23.1",
-    "uglifyjs-webpack-plugin": "^2.1.3",
-    "url-loader": "^2.0.1",
-    "webpack": "^4.46.0",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.12",
+    "style-loader": "^3.3.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.58.1",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.0",
     "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^4.2.1"
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@mui/material": "^5.0.0",
-    "clsx": "^1.0.4",
-    "lodash": "^4.17.14",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "country-flag-icons": "^1.4.10",
+    "lodash": "^4.17.21"
   },
   "peerDependencies": {
-    "@emotion/react": "^11.4.0",
+    "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
     "@mui/material": "^5.0.0",
+    "clsx": "^1.1.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
