@@ -1,18 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import MenuItem from '@mui/material/MenuItem';
+import Flags from 'country-flag-icons/react/3x2';
 
 class Item extends React.PureComponent {
   render() {
     const ref = React.createRef();
     const {
-      name,
-      iso2,
-      dialCode,
-      localization,
-      itemRef,
-      native,
-      ...restProps
+      name, iso2, dialCode, localization,
+      itemRef, native, className = '', ...restProps
     } = this.props;
 
     if (native) {
@@ -24,10 +20,14 @@ class Item extends React.PureComponent {
           value={iso2}
           {...restProps}
         >
-          {localization || name} {`+${dialCode}`}
+          {localization || name}
+          {' '}
+          {`+${dialCode}`}
         </option>
       );
     }
+
+    const FlagComponent = Flags[iso2.toUpperCase()];
 
     return (
       <MenuItem
@@ -37,9 +37,15 @@ class Item extends React.PureComponent {
         data-country-code={iso2}
         {...restProps}
       >
-        <div className={`flag ${iso2} margin`} />
+        {Boolean(FlagComponent) && (
+          <div className={className}>
+            <FlagComponent />
+          </div>
+        )}
 
-        <span className="country-name">{localization || name}</span>
+        <span className="country-name">
+          {localization || name}
+        </span>
 
         <span className="dial-code">{`+${dialCode}`}</span>
       </MenuItem>
