@@ -1,97 +1,61 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import clsx from 'clsx';
+import Flags from 'country-flag-icons/react/3x2'
 import TextField from '@mui/material/TextField';
 import InputAdornment from '@mui/material/InputAdornment';
-import Button from '@mui/material/Button';
+import IconButton from '@mui/material/IconButton';
 import Menu from '@mui/material/Menu';
 import Divider from '@mui/material/Divider';
 import NativeSelect from '@mui/material/NativeSelect';
 import {
-  some,
-  find,
-  reduce,
-  map,
-  filter,
-  includes,
-  findIndex,
-  head,
-  tail,
-  debounce,
-  memoize,
-  trim,
-  startsWith,
-  isString,
+  some, find, reduce, map, filter, includes, findIndex,
+  head, tail, debounce, memoize, trim, startsWith, isString,
 } from 'lodash';
 import countryData from '../country_data';
 import Item from './Item';
-import '../styles.less';
-import '../flags.png';
 
 class MaterialUiPhoneNumber extends React.Component {
   flags = {};
 
-  guessSelectedCountry = memoize(
-    (inputNumber, onlyCountries, defaultCountry) => {
-      const secondBestGuess = find(onlyCountries, { iso2: defaultCountry }) || {};
-      if (trim(inputNumber) === '') return secondBestGuess;
+  guessSelectedCountry = memoize((inputNumber, onlyCountries, defaultCountry) => {
+    const secondBestGuess = find(onlyCountries, { iso2: defaultCountry }) || {};
+    if (trim(inputNumber) === '') return secondBestGuess;
 
-      const bestGuess = reduce(
-        onlyCountries,
-        (selectedCountry, country) => {
-          if (startsWith(inputNumber, country.dialCode)) {
-            if (country.dialCode.length > selectedCountry.dialCode.length) {
-              return country;
-            }
-            if (
-              country.dialCode.length === selectedCountry.dialCode.length
-              && country.priority < selectedCountry.priority
-            ) {
-              return country;
-            }
-          }
-          return selectedCountry;
-        },
-        { dialCode: '', priority: 10001 },
-        this,
-      );
+    const bestGuess = reduce(onlyCountries, (selectedCountry, country) => {
+      if (startsWith(inputNumber, country.dialCode)) {
+        if (country.dialCode.length > selectedCountry.dialCode.length) {
+          return country;
+        }
+        if (country.dialCode.length === selectedCountry.dialCode.length && country.priority < selectedCountry.priority) {
+          return country;
+        }
+      }
+      return selectedCountry;
+    }, { dialCode: '', priority: 10001 }, this);
 
-      if (!bestGuess.name) return secondBestGuess;
-      return bestGuess;
-    },
-  );
+    if (!bestGuess.name) return secondBestGuess;
+    return bestGuess;
+  });
 
   constructor(props) {
     super(props);
     let filteredCountries = countryData.allCountries;
 
-    if (props.disableAreaCodes) {
-      filteredCountries = this.deleteAreaCodes(filteredCountries);
-    }
-    if (props.regions) {
-      filteredCountries = this.filterRegions(props.regions, filteredCountries);
-    }
+    if (props.disableAreaCodes) filteredCountries = this.deleteAreaCodes(filteredCountries);
+    if (props.regions) filteredCountries = this.filterRegions(props.regions, filteredCountries);
 
     const onlyCountries = this.excludeCountries(
-      this.getOnlyCountries(props.onlyCountries, filteredCountries),
-      props.excludeCountries,
+      this.getOnlyCountries(props.onlyCountries, filteredCountries), props.excludeCountries,
     );
 
-    const preferredCountries = filter(filteredCountries, (country) => some(
-      props.preferredCountries,
-      (preferredCountry) => preferredCountry === country.iso2,
-    ));
+    const preferredCountries = filter(filteredCountries, (country) => some(props.preferredCountries, (preferredCountry) => preferredCountry === country.iso2));
 
     const inputNumber = props.value || '';
 
     let countryGuess;
     if (inputNumber.length > 1) {
       // Country detect by value field
-      countryGuess = this.guessSelectedCountry(
-        inputNumber.replace(/\D/g, '').substring(0, 6),
-        onlyCountries,
-        props.defaultCountry,
-      ) || 0;
+      countryGuess = this.guessSelectedCountry(inputNumber.replace(/\D/g, '').substring(0, 6), onlyCountries, props.defaultCountry) || 0;
     } else if (props.defaultCountry) {
       // Default country
       countryGuess = find(onlyCountries, { iso2: props.defaultCountry }) || 0;
@@ -101,17 +65,15 @@ class MaterialUiPhoneNumber extends React.Component {
     }
 
     const countryGuessIndex = findIndex(this.allCountries, countryGuess);
-    const dialCode = inputNumber.length < 2
+    const dialCode = (
+      inputNumber.length < 2
       && countryGuess
       && !startsWith(inputNumber.replace(/\D/g, ''), countryGuess.dialCode)
-      ? countryGuess.dialCode
-      : '';
+    ) ? countryGuess.dialCode : '';
 
-    const formattedNumber = inputNumber === '' && countryGuess === 0
-      ? ''
+    const formattedNumber = (inputNumber === '' && countryGuess === 0) ? ''
       : this.formatNumber(
-        (props.disableCountryCode ? '' : dialCode)
-              + inputNumber.replace(/\D/g, ''),
+        (props.disableCountryCode ? '' : dialCode) + inputNumber.replace(/\D/g, ''),
         countryGuess.name ? countryGuess.format : undefined,
       );
 
@@ -144,11 +106,7 @@ class MaterialUiPhoneNumber extends React.Component {
       this.updateDefaultCountry(defaultCountry);
     }
 
-    if (
-      typeof value === 'string'
-      && value !== prevValue
-      && value !== formattedNumber
-    ) {
+    if (typeof value === 'string' && value !== prevValue && value !== formattedNumber) {
       this.updateFormattedNumber(value);
     }
   }
@@ -167,11 +125,7 @@ class MaterialUiPhoneNumber extends React.Component {
     const { onlyCountries } = this.state;
 
     // don't include the preferred countries in search
-    const probableCountries = filter(
-      onlyCountries,
-      (country) => startsWith(country.name.toLowerCase(), queryString.toLowerCase()),
-      this,
-    );
+    const probableCountries = filter(onlyCountries, (country) => startsWith(country.name.toLowerCase(), queryString.toLowerCase()), this);
     return probableCountries[0];
   });
 
@@ -179,17 +133,14 @@ class MaterialUiPhoneNumber extends React.Component {
     if (onlyCountriesArray.length === 0) return filteredCountries;
 
     return filteredCountries.filter((country) => onlyCountriesArray.some((element) => element === country.iso2));
-  };
+  }
 
   excludeCountries = (selectedCountries, excludedCountries) => {
     if (excludedCountries.length === 0) {
       return selectedCountries;
     }
-    return filter(
-      selectedCountries,
-      (selCountry) => !includes(excludedCountries, selCountry.iso2),
-    );
-  };
+    return filter(selectedCountries, (selCountry) => !includes(excludedCountries, selCountry.iso2));
+  }
 
   filterRegions = (regions, filteredCountries) => {
     if (typeof regions === 'string') {
@@ -201,7 +152,7 @@ class MaterialUiPhoneNumber extends React.Component {
       const matches = regions.map((region) => country.regions.some((element) => element === region));
       return matches.some((el) => el);
     });
-  };
+  }
 
   // Countries array methods
   deleteAreaCodes = (filteredCountries) => filteredCountries.filter((country) => country.isAreaCode !== true);
@@ -216,25 +167,19 @@ class MaterialUiPhoneNumber extends React.Component {
     this.setState({
       defaultCountry: country,
       selectedCountry: newSelectedCountry,
-      formattedNumber: disableCountryCode
-        ? ''
-        : `+${newSelectedCountry.dialCode}`,
+      formattedNumber: disableCountryCode ? '' : `+${newSelectedCountry.dialCode}`,
     });
-  };
+  }
 
   // View methods
   scrollTo = (country) => {
-    if (!country) {
-      return;
-    }
+    if (!country) { return; }
 
     const container = this.dropdownContainerRef;
 
-    if (!container || !document.body) {
-      return;
-    }
+    if (!container || !document.body) { return; }
     container.scrollTop = country.offsetTop;
-  };
+  }
 
   formatNumber = (text, patternArg) => {
     const { disableCountryCode, enableLongNumbers, autoFormat } = this.props;
@@ -258,30 +203,26 @@ class MaterialUiPhoneNumber extends React.Component {
       return disableCountryCode ? text : `+${text}`;
     }
 
-    const formattedObject = reduce(
-      pattern,
-      (acc, character) => {
-        if (acc.remainingText.length === 0) {
-          return acc;
-        }
+    const formattedObject = reduce(pattern, (acc, character) => {
+      if (acc.remainingText.length === 0) {
+        return acc;
+      }
 
-        if (character !== '.') {
-          return {
-            formattedText: acc.formattedText + character,
-            remainingText: acc.remainingText,
-          };
-        }
-
+      if (character !== '.') {
         return {
-          formattedText: acc.formattedText + head(acc.remainingText),
-          remainingText: tail(acc.remainingText),
+          formattedText: acc.formattedText + character,
+          remainingText: acc.remainingText,
         };
-      },
-      {
-        formattedText: '',
-        remainingText: text.split(''),
-      },
-    );
+      }
+
+      return {
+        formattedText: acc.formattedText + head(acc.remainingText),
+        remainingText: tail(acc.remainingText),
+      };
+    }, {
+      formattedText: '',
+      remainingText: text.split(''),
+    });
 
     let formattedNumber;
     if (enableLongNumbers) {
@@ -291,11 +232,9 @@ class MaterialUiPhoneNumber extends React.Component {
     }
 
     // Always close brackets
-    if (formattedNumber.includes('(') && !formattedNumber.includes(')')) {
-      formattedNumber += ')';
-    }
+    if (formattedNumber.includes('(') && !formattedNumber.includes(')')) formattedNumber += ')';
     return formattedNumber;
-  };
+  }
 
   // Put the cursor to the end of the input (usually after a focus event)
   cursorToEnd = () => {
@@ -309,9 +248,9 @@ class MaterialUiPhoneNumber extends React.Component {
         input.setSelectionRange(len, len);
       }
     }
-  };
+  }
 
-  getElement = (index) => this.flags[`flag_no_${index}`];
+  getElement = (index) => this.flags[`flag_no_${index}`]
 
   // return country data from state
   getCountryData = () => {
@@ -324,21 +263,15 @@ class MaterialUiPhoneNumber extends React.Component {
       dialCode: selectedCountry.dialCode || '',
       countryCode: selectedCountry.iso2 || '',
     };
-  };
+  }
 
   handleInput = (e) => {
     let { selectedCountry: newSelectedCountry, freezeSelection } = this.state;
     const {
-      selectedCountry,
-      formattedNumber: oldFormattedText,
-      onlyCountries,
-      defaultCountry,
+      selectedCountry, formattedNumber: oldFormattedText, onlyCountries, defaultCountry,
     } = this.state;
     const {
-      disableCountryCode,
-      countryCodeEditable,
-      isModernBrowser,
-      onChange,
+      disableCountryCode, countryCodeEditable, isModernBrowser, onChange,
     } = this.props;
 
     let formattedNumber = disableCountryCode ? '' : '+';
@@ -374,62 +307,43 @@ class MaterialUiPhoneNumber extends React.Component {
       // we don't need to send the whole number to guess the country... only the first 6 characters are enough
       // the guess country function can then use memoization much more effectively since the set of input it
       // gets has drastically reduced
-      if (
-        !freezeSelection
-        || selectedCountry.dialCode.length > inputNumber.length
-      ) {
-        newSelectedCountry = this.guessSelectedCountry(
-          inputNumber.substring(0, 6),
-          onlyCountries,
-          defaultCountry,
-        );
+      if (!freezeSelection || selectedCountry.dialCode.length > inputNumber.length) {
+        newSelectedCountry = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, defaultCountry);
         freezeSelection = false;
       }
       // let us remove all non numerals from the input
-      formattedNumber = this.formatNumber(
-        inputNumber,
-        newSelectedCountry.format,
-      );
+      formattedNumber = this.formatNumber(inputNumber, newSelectedCountry.format);
     }
 
     let caretPosition = e.target.selectionStart;
     const diff = formattedNumber.length - oldFormattedText.length;
 
-    this.setState(
-      {
-        formattedNumber,
-        freezeSelection,
-        selectedCountry: newSelectedCountry.dialCode
-          ? newSelectedCountry
-          : selectedCountry,
-      },
-      () => {
-        if (isModernBrowser) {
-          if (diff > 0) {
-            caretPosition -= diff;
-          }
-
-          const lastChar = formattedNumber.charAt(formattedNumber.length - 1);
-
-          if (lastChar === ')') {
-            this.inputRef.setSelectionRange(
-              formattedNumber.length - 1,
-              formattedNumber.length - 1,
-            );
-          } else if (
-            caretPosition > 0
-            && oldFormattedText.length >= formattedNumber.length
-          ) {
-            this.inputRef.setSelectionRange(caretPosition, caretPosition);
-          }
+    this.setState({
+      formattedNumber,
+      freezeSelection,
+      selectedCountry: newSelectedCountry.dialCode
+        ? newSelectedCountry
+        : selectedCountry,
+    }, () => {
+      if (isModernBrowser) {
+        if (diff > 0) {
+          caretPosition -= diff;
         }
 
-        if (onChange) {
-          onChange(formattedNumber, this.getCountryData());
+        const lastChar = formattedNumber.charAt(formattedNumber.length - 1);
+
+        if (lastChar === ')') {
+          this.inputRef.setSelectionRange(formattedNumber.length - 1, formattedNumber.length - 1);
+        } else if (caretPosition > 0 && oldFormattedText.length >= formattedNumber.length) {
+          this.inputRef.setSelectionRange(caretPosition, caretPosition);
         }
-      },
-    );
-  };
+      }
+
+      if (onChange) {
+        onChange(formattedNumber, this.getCountryData());
+      }
+    });
+  }
 
   handleRefInput = (ref) => {
     const { inputRef, InputProps } = this.props;
@@ -458,49 +372,32 @@ class MaterialUiPhoneNumber extends React.Component {
     if (onClick) {
       onClick(e, this.getCountryData());
     }
-  };
+  }
 
   handleFlagItemClick = (country) => {
     const { formattedNumber, selectedCountry, onlyCountries } = this.state;
     const { onChange } = this.props;
 
     const currentSelectedCountry = selectedCountry;
-    const nextSelectedCountry = isString(country)
-      ? find(onlyCountries, (countryItem) => countryItem.iso2 === country)
-      : find(onlyCountries, country);
+    const nextSelectedCountry = isString(country) ? find(onlyCountries, (countryItem) => countryItem.iso2 === country) : find(onlyCountries, country);
 
-    const unformattedNumber = formattedNumber
-      .replace(' ', '')
-      .replace('(', '')
-      .replace(')', '')
-      .replace('-', '');
-    const newNumber = unformattedNumber.length > 1
-      ? unformattedNumber.replace(
-        currentSelectedCountry.dialCode,
-        nextSelectedCountry.dialCode,
-      )
-      : nextSelectedCountry.dialCode;
+    const unformattedNumber = formattedNumber.replace(' ', '').replace('(', '').replace(')', '').replace('-', '');
+    const newNumber = unformattedNumber.length > 1 ? unformattedNumber.replace(currentSelectedCountry.dialCode, nextSelectedCountry.dialCode) : nextSelectedCountry.dialCode;
 
-    const newFormattedNumber = this.formatNumber(
-      newNumber.replace(/\D/g, ''),
-      nextSelectedCountry.format,
-    );
+    const newFormattedNumber = this.formatNumber(newNumber.replace(/\D/g, ''), nextSelectedCountry.format);
 
-    this.setState(
-      {
-        anchorEl: null,
-        selectedCountry: nextSelectedCountry,
-        freezeSelection: true,
-        formattedNumber: newFormattedNumber,
-      },
-      () => {
-        this.cursorToEnd();
-        if (onChange) {
-          onChange(newFormattedNumber, this.getCountryData());
-        }
-      },
-    );
-  };
+    this.setState({
+      anchorEl: null,
+      selectedCountry: nextSelectedCountry,
+      freezeSelection: true,
+      formattedNumber: newFormattedNumber,
+    }, () => {
+      this.cursorToEnd();
+      if (onChange) {
+        onChange(newFormattedNumber, this.getCountryData());
+      }
+    });
+  }
 
   handleInputFocus = (e) => {
     const { selectedCountry } = this.state;
@@ -508,17 +405,10 @@ class MaterialUiPhoneNumber extends React.Component {
 
     // if the input is blank, insert dial code of the selected country
     if (this.inputRef) {
-      if (
-        this.inputRef.value === '+'
-        && selectedCountry
-        && !disableCountryCode
-      ) {
-        this.setState(
-          {
-            formattedNumber: `+${selectedCountry.dialCode}`,
-          },
-          () => setTimeout(this.cursorToEnd, 10),
-        );
+      if (this.inputRef.value === '+' && selectedCountry && !disableCountryCode) {
+        this.setState({
+          formattedNumber: `+${selectedCountry.dialCode}`,
+        }, () => setTimeout(this.cursorToEnd, 10));
       }
     }
 
@@ -529,7 +419,7 @@ class MaterialUiPhoneNumber extends React.Component {
     }
 
     setTimeout(this.cursorToEnd, 10);
-  };
+  }
 
   handleInputBlur = (e) => {
     const { placeholder, onBlur } = this.props;
@@ -541,27 +431,20 @@ class MaterialUiPhoneNumber extends React.Component {
     if (onBlur) {
       onBlur(e, this.getCountryData());
     }
-  };
+  }
 
   getHighlightCountryIndex = (direction) => {
-    const {
-      highlightCountryIndex: oldHighlightCountryIndex,
-      onlyCountries,
-      preferredCountries,
-    } = this.state;
+    const { highlightCountryIndex: oldHighlightCountryIndex, onlyCountries, preferredCountries } = this.state;
 
     // had to write own function because underscore does not have findIndex. lodash has it
     const highlightCountryIndex = oldHighlightCountryIndex + direction;
 
-    if (
-      highlightCountryIndex < 0
-      || highlightCountryIndex >= onlyCountries.length + preferredCountries.length
-    ) {
+    if (highlightCountryIndex < 0 || highlightCountryIndex >= (onlyCountries.length + preferredCountries.length)) {
       return highlightCountryIndex - direction;
     }
 
     return highlightCountryIndex;
-  };
+  }
 
   searchCountry = () => {
     const { queryString, onlyCountries, preferredCountries } = this.state;
@@ -571,20 +454,13 @@ class MaterialUiPhoneNumber extends React.Component {
 
     this.scrollTo(this.getElement(probableCandidateIndex), true);
 
-    this.setState({
-      queryString: '',
-      highlightCountryIndex: probableCandidateIndex,
-    });
-  };
+    this.setState({ queryString: '', highlightCountryIndex: probableCandidateIndex });
+  }
 
   handleKeydown = (e) => {
     const {
-      anchorEl,
-      highlightCountryIndex,
-      preferredCountries,
-      onlyCountries,
-      queryString,
-      debouncedQueryStingSearcher,
+      anchorEl, highlightCountryIndex, preferredCountries, onlyCountries,
+      queryString, debouncedQueryStingSearcher,
     } = this.state;
     const { keys, disabled } = this.props;
 
@@ -598,17 +474,13 @@ class MaterialUiPhoneNumber extends React.Component {
     }
 
     const moveHighlight = (direction) => {
-      this.setState(
-        {
-          highlightCountryIndex: this.getHighlightCountryIndex(direction),
-        },
-        () => {
-          this.scrollTo(
-            this.getElement(highlightCountryIndex + preferredCountries.length),
-            true,
-          );
-        },
-      );
+      this.setState({
+        highlightCountryIndex: this.getHighlightCountryIndex(direction),
+      }, () => {
+        this.scrollTo(this.getElement(
+          highlightCountryIndex + preferredCountries.length,
+        ), true);
+      });
     };
 
     switch (e.which) {
@@ -622,27 +494,18 @@ class MaterialUiPhoneNumber extends React.Component {
         this.handleFlagItemClick(onlyCountries[highlightCountryIndex], e);
         break;
       case keys.ESC:
-        this.setState(
-          {
-            anchorEl: null,
-          },
-          this.cursorToEnd,
-        );
+        this.setState({
+          anchorEl: null,
+        }, this.cursorToEnd);
         break;
       default:
-        if (
-          (e.which >= keys.A && e.which <= keys.Z)
-          || e.which === keys.SPACE
-        ) {
-          this.setState(
-            {
-              queryString: queryString + String.fromCharCode(e.which),
-            },
-            debouncedQueryStingSearcher,
-          );
+        if ((e.which >= keys.A && e.which <= keys.Z) || e.which === keys.SPACE) {
+          this.setState({
+            queryString: queryString + String.fromCharCode(e.which),
+          }, debouncedQueryStingSearcher);
         }
     }
-  };
+  }
 
   handleInputKeyDown = (e) => {
     const { keys, onEnterKeyPress, onKeyDown } = this.props;
@@ -653,7 +516,7 @@ class MaterialUiPhoneNumber extends React.Component {
     if (onKeyDown) {
       onKeyDown(e);
     }
-  };
+  }
 
   checkIfValid = () => {
     const { formattedNumber } = this.state;
@@ -674,21 +537,14 @@ class MaterialUiPhoneNumber extends React.Component {
     // otherwise use logic for finding country based on country prefix.
     if (!inputNumber.startsWith('+')) {
       countryGuess = find(onlyCountries, { iso2: defaultCountry });
-      const dialCode = countryGuess
-        && !startsWith(inputNumber.replace(/\D/g, ''), countryGuess.dialCode)
-        ? countryGuess.dialCode
-        : '';
+      const dialCode = countryGuess && !startsWith(inputNumber.replace(/\D/g, ''), countryGuess.dialCode) ? countryGuess.dialCode : '';
       formattedNumber = this.formatNumber(
         (disableCountryCode ? '' : dialCode) + inputNumber.replace(/\D/g, ''),
         countryGuess ? countryGuess.format : undefined,
       );
     } else {
       inputNumber = inputNumber.replace(/\D/g, '');
-      countryGuess = this.guessSelectedCountry(
-        inputNumber.substring(0, 6),
-        onlyCountries,
-        defaultCountry,
-      );
+      countryGuess = this.guessSelectedCountry(inputNumber.substring(0, 6), onlyCountries, defaultCountry);
       formattedNumber = this.formatNumber(inputNumber, countryGuess.format);
     }
 
@@ -728,9 +584,9 @@ class MaterialUiPhoneNumber extends React.Component {
     };
 
     const {
-      classes, dropdownClass, localization, disableDropdown, native,
+      classes, dropdownClass, localization, disableDropdown,
+      native,
     } = this.props;
-    const inputFlagClasses = `flag ${selectedCountry.iso2}`;
 
     onlyCountries.sort((a, b) => {
       const localizedA = localization[a.name] || a.name;
@@ -740,74 +596,70 @@ class MaterialUiPhoneNumber extends React.Component {
 
     const isSelected = (country) => Boolean(selectedCountry && selectedCountry.dialCode === country.dialCode);
 
-    const dropdownProps = disableDropdown
-      ? {}
-      : {
-        startAdornment: (
-          <InputAdornment className={styles.positionStart} position="start">
-            {native ? (
-              <>
-                <NativeSelect
-                  id="country-menu"
-                  open={Boolean(anchorEl)}
-                  onClose={() => this.setState({ anchorEl: null })}
-                  className={styles.native}
-                  classes={{
-                    root: clsx(
-                      styles.nativeRoot,
-                      'native',
-                      inputFlagClasses,
-                    ),
-                    select: styles.nativeSelect,
-                  }}
-                  onChange={(e) => this.handleFlagItemClick(e.target.value)}
-                  disableUnderline
-                >
-                  {!!preferredCountries.length
-                      && map(preferredCountries, (country, index) => (
-                        <Item
-                          key={`preferred_${country.iso2}_${index}`}
-                          itemRef={(node) => {
-                            this.flags[`flag_no_${index}`] = node;
-                          }}
-                          name={country.name}
-                          iso2={country.iso2}
-                          dialCode={country.dialCode}
-                          localization={
-                            localization && localization[country.name]
-                          }
-                          native
-                        />
-                      ))}
+    const FlagComponent = Flags[selectedCountry.iso2.toUpperCase()];
 
-                  {map(onlyCountries, (country, index) => (
-                    <Item
-                      key={`preferred_${country.iso2}_${index}`}
-                      itemRef={(node) => {
-                        this.flags[`flag_no_${index}`] = node;
-                      }}
-                      name={country.name}
-                      iso2={country.iso2}
-                      dialCode={country.dialCode}
-                      localization={
-                          localization && localization[country.name]
-                        }
-                      native
-                    />
-                  ))}
-                </NativeSelect>
-              </>
-            ) : (
+    const dropdownProps = disableDropdown ? {} : {
+      startAdornment: (
+        <InputAdornment
+          style={styles.positionStart}
+          position="start"
+        >
+          {native ? (
+            <>
+              <NativeSelect
+                id="country-menu"
+                open={Boolean(anchorEl)}
+                onClose={() => this.setState({ anchorEl: null })}
+                style={styles.native}
+                sx={{
+                  root: styles.nativeRoot, // clsx(classes.nativeRoot, 'native'),
+                  select: styles.nativeSelect,
+                }}
+                onChange={(e) => this.handleFlagItemClick(e.target.value)}
+                IconComponent={Boolean(FlagComponent) && FlagComponent}
+                disableUnderline
+              >
+                {!!preferredCountries.length && map(preferredCountries, (country, index) => (
+                  <Item
+                    key={`preferred_${country.iso2}_${index}`}
+                    itemRef={(node) => {
+                      this.flags[`flag_no_${index}`] = node;
+                    }}
+                    name={country.name}
+                    iso2={country.iso2}
+                    dialCode={country.dialCode}
+                    localization={localization && localization[country.name]}
+                    native
+                  />
+                ))}
+
+                {map(onlyCountries, (country, index) => (
+                  <Item
+                    key={`preferred_${country.iso2}_${index}`}
+                    itemRef={(node) => {
+                      this.flags[`flag_no_${index}`] = node;
+                    }}
+                    name={country.name}
+                    iso2={country.iso2}
+                    dialCode={country.dialCode}
+                    localization={localization && localization[country.name]}
+                    native
+                  />
+                ))}
+              </NativeSelect>
+            </>
+          )
+            : (
               <>
-                <Button
-                  className={styles.flagButton}
+                <IconButton
+                  style={styles.flagButton}
                   aria-owns={anchorEl ? 'country-menu' : null}
                   aria-label="Select country"
                   onClick={(e) => this.setState({ anchorEl: e.currentTarget })}
                   aria-haspopup
                 >
-                  <div className={inputFlagClasses} />
-                </Button>
+                  {Boolean(FlagComponent) && <FlagComponent className="margin" />}
+                </IconButton>
 
                 <Menu
                   className={dropdownClass}
@@ -816,23 +668,21 @@ class MaterialUiPhoneNumber extends React.Component {
                   open={Boolean(anchorEl)}
                   onClose={() => this.setState({ anchorEl: null })}
                 >
-                  {!!preferredCountries.length
-                      && map(preferredCountries, (country, index) => (
-                        <Item
-                          key={`preferred_${country.iso2}_${index}`}
-                          itemRef={(node) => {
-                            this.flags[`flag_no_${index}`] = node;
-                          }}
-                          selected={isSelected(country)}
-                          onClick={() => this.handleFlagItemClick(country)}
-                          name={country.name}
-                          iso2={country.iso2}
-                          dialCode={country.dialCode}
-                          localization={
-                            localization && localization[country.name]
-                          }
-                        />
-                      ))}
+                  {!!preferredCountries.length && map(preferredCountries, (country, index) => (
+                    <Item
+                      key={`preferred_${country.iso2}_${index}`}
+                      itemRef={(node) => {
+                        this.flags[`flag_no_${index}`] = node;
+                      }}
+                      selected={isSelected(country)}
+                      onClick={() => this.handleFlagItemClick(country)}
+                      name={country.name}
+                      iso2={country.iso2}
+                      dialCode={country.dialCode}
+                      localization={localization && localization[country.name]}
+                      style={styles.flagIcon}
+                    />
+                  ))}
 
                   {!!preferredCountries.length && <Divider />}
 
@@ -847,52 +697,34 @@ class MaterialUiPhoneNumber extends React.Component {
                       name={country.name}
                       iso2={country.iso2}
                       dialCode={country.dialCode}
-                      localization={
-                          localization && localization[country.name]
-                        }
+                      localization={localization && localization[country.name]}
+                      style={styles.flagIcon}
                     />
                   ))}
                 </Menu>
               </>
             )}
-          </InputAdornment>
-        ),
-      };
+        </InputAdornment>
+      ),
+    };
 
     return dropdownProps;
   };
 
   render() {
-    const { formattedNumber, placeholder: statePlaceholder } = this.state;
+    const {
+      formattedNumber, placeholder: statePlaceholder,
+    } = this.state;
 
     const {
       // start placeholder props
-      native,
-      defaultCountry,
-      excludeCountries,
-      onlyCountries,
-      preferredCountries,
-      dropdownClass,
-      autoFormat,
-      disableAreaCodes,
-      isValid,
-      disableCountryCode,
-      disableDropdown,
-      enableLongNumbers,
-      countryCodeEditable,
-      onEnterKeyPress,
-      isModernBrowser,
-      classes,
-      keys,
-      localization,
-      placeholder,
-      regions,
-      onChange,
+      native, defaultCountry, excludeCountries, onlyCountries, preferredCountries,
+      dropdownClass, autoFormat, disableAreaCodes, isValid, disableCountryCode,
+      disableDropdown, enableLongNumbers, countryCodeEditable, onEnterKeyPress,
+      isModernBrowser, classes, keys, localization, placeholder, regions, onChange,
       value,
       // end placeholder props
-      inputClass,
-      error,
-      InputProps,
+      inputClass, error, InputProps,
       ...restProps
     } = this.props;
 
@@ -985,11 +817,7 @@ MaterialUiPhoneNumber.defaultProps = {
 
   autoFormat: true,
   disableAreaCodes: false,
-  isValid: (inputNumber) => some(
-    countryData.allCountries,
-    (country) => startsWith(inputNumber, country.dialCode)
-        || startsWith(country.dialCode, inputNumber),
-  ),
+  isValid: (inputNumber) => some(countryData.allCountries, (country) => startsWith(inputNumber, country.dialCode) || startsWith(country.dialCode, inputNumber)),
   disableCountryCode: false,
   disableDropdown: false,
   enableLongNumbers: false,
@@ -999,12 +827,10 @@ MaterialUiPhoneNumber.defaultProps = {
 
   localization: {},
 
-  onEnterKeyPress: () => {},
-  onChange: () => {},
+  onEnterKeyPress: () => { },
+  onChange: () => { },
 
-  isModernBrowser: () => (document.createElement
-    ? Boolean(document.createElement('input').setSelectionRange)
-    : false),
+  isModernBrowser: () => (document.createElement ? Boolean(document.createElement('input').setSelectionRange) : false),
 
   keys: {
     UP: 38,
