@@ -5,7 +5,7 @@ const ROOT_PATH = path.resolve(__dirname);
 // Common configuration settings
 module.exports = {
   entry: {
-    'material-ui-phone-number': path.resolve(ROOT_PATH, 'src/index.js'),
+    'mui-phone-number': path.resolve(ROOT_PATH, 'src/index.js'),
   },
   resolve: {
     extensions: ['.js', '.jsx'],
@@ -14,11 +14,10 @@ module.exports = {
   output: {
     path: path.resolve(ROOT_PATH, 'dist'),
     filename: 'index.js',
-    library: 'MaterialUiPhoneNumber',
+    library: 'MUiPhoneNumber',
     umdNamedDefine: true,
     globalObject: `(typeof self !== 'undefined' ? self : this)`, // eslint-disable-line
     libraryTarget: 'umd',
-    futureEmitAssets: true,
   },
   module: {
     rules: [
@@ -41,9 +40,8 @@ module.exports = {
         },
       },
       {
-        test: /\.png.*$/,
-        use: ['url-loader?limit=100000&mimetype=image/png'],
-        exclude: /node_modules/,
+        test: /\.(gif|png|jpe?g|svg)$/i,
+        type: 'asset/resource',
       },
       {
         test: /\.less$/,
